+++ webui/src/components/form/index.ts
@@ -1,6 +1,5 @@
 export {
   Button,
-  Badge,
   FormActions,
   FormError,
   FormField,
