+++ webui/src/routes/import/-ui/singles-import-tab.tsx
@@ -1,8 +1,8 @@
 import clsx from 'clsx';
 import { useEffect } from 'react';
 
-import { Badge, Button, Checkbox, TextInput } from '@/components/form/form';
-import { Notice } from '@/components/primitives';
+import { Button, Checkbox, TextInput } from '@/components/form/form';
+import { Badge, Notice } from '@/components/primitives';
 
 import type { SingleSearchState } from '../-import.store';
 import type { ImportTrackResult } from '../-import.types';
